+++ src/containerapp/azext_containerapp/commands.py
@@ -6,9 +6,10 @@
 # pylint: disable=line-too-long, too-many-statements, bare-except
 # from azure.cli.core.commands import CliCommandType
 # from azure.mgmt.core.tools import is_valid_resource_id, parse_resource_id
-from azure.cli.command_modules.containerapp._transformers import (transform_containerapp_output, transform_containerapp_list_output)
+from azure.cli.command_modules.containerapp._transformers import transform_containerapp_output
 from azext_containerapp._client_factory import ex_handler_factory
 from ._transformers import (transform_sensitive_values,
+                            transform_containerapp_list_output_with_kind,
                             transform_telemetry_data_dog_values,
                             transform_telemetry_app_insights_values,
                             transform_telemetry_otlp_values,
@@ -25,7 +26,7 @@ from ._validators import validate_debug
 def load_command_table(self, args):
     with self.command_group('containerapp') as g:
         g.custom_show_command('show', 'show_containerapp', table_transformer=transform_containerapp_output)
-        g.custom_command('list', 'list_containerapp', table_transformer=transform_containerapp_list_output)
+        g.custom_command('list', 'list_containerapp', table_transformer=transform_containerapp_list_output_with_kind)
         g.custom_command('create', 'create_containerapp', supports_no_wait=True, exception_handler=ex_handler_factory(), table_transformer=transform_containerapp_output, transform=transform_sensitive_values)
         g.custom_command('update', 'update_containerapp', supports_no_wait=True, exception_handler=ex_handler_factory(), table_transformer=transform_containerapp_output, transform=transform_sensitive_values)
         g.custom_command('delete', 'delete_containerapp', supports_no_wait=True, confirmation=True, exception_handler=ex_handler_factory())