+++ src/containerapp/azext_containerapp/custom.py
@@ -705,7 +705,7 @@ def show_containerapp(cmd, name, resource_group_name, show_secrets=False):
     return containerapp_base_decorator.show()
 
 
-def list_containerapp(cmd, resource_group_name=None, managed_env=None, environment_type="all"):
+def list_containerapp(cmd, resource_group_name=None, managed_env=None, environment_type="all", kind=None):
     raw_parameters = locals()
     containerapp_list_decorator = ContainerAppPreviewListDecorator(
         cmd=cmd,