+++ src/containerapp/HISTORY.rst
@@ -6,6 +6,8 @@ upcoming
 ++++++
 * 'az containerapp function invocations': Update application insights query 
 * 'az containerapp function keys': Update minimum replica check 
+* 'az containerapp list': Add `--kind` parameter to filter container apps by kind 
+* 'az containerapp function invocations': Fix issue when cloud role name is set
 
 1.3.0b1
 ++++++